Move the driver package into the internal package
